requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
apscheduler
httpx
requests
beautifulsoup4
lxml
selenium
fake-useragent
tenacity
python-dotenv
pydantic-settings
transformers
torch
scikit-learn
numpy
pandas
loguru
openpyxl